+++ jwst/assign_wcs/niriss.py
@@ -31,7 +31,7 @@ def create_pipeline(input_model, reference_files):
 
     Parameters
     ----------
-    input_model : `~jwst.datamodel.DataModel`
+    input_model : `~jwst.datamodel.JwstDataModel`
         Input datamodel for processing
     reference_files : dict
         The dictionary of reference file names and their associated files
@@ -117,7 +117,7 @@ def niriss_soss(input_model, reference_files):
 
     Parameters
     ----------
-    input_model : `~jwst.datamodel.DataModel`
+    input_model : `~jwst.datamodel.JwstDataModel`
         Input datamodel for processing
     reference_files : dict
         The dictionary of reference file names and their associated files
@@ -216,7 +216,7 @@ def imaging(input_model, reference_files):
 
     Parameters
     ----------
-    input_model : `~jwst.datamodel.DataModel`
+    input_model : `~jwst.datamodel.JwstDataModel`
         Input datamodel for processing
     reference_files : dict
         The dictionary of reference file names and their associated files
@@ -268,7 +268,7 @@ def imaging_distortion(input_model, reference_files):
 
     Parameters
     ----------
-    input_model : `~jwst.datamodel.DataModel`
+    input_model : `~jwst.datamodel.JwstDataModel`
         Input datamodel for processing
     reference_files : dict
         The dictionary of reference file names and their associated files.