+++ jwst/pipeline/calwebb_spec3.py
@@ -73,7 +73,7 @@ class Spec3Pipeline(Pipeline):
 
         Parameters
         ----------
-        input: str, Level3 Association, or ~jwst.datamodels.DataModel
+        input: str, Level3 Association, or ~jwst.datamodels.JwstDataModel
             The exposure or association of exposures to process
         """
         self.log.info('Starting calwebb_spec3 ...')